♻️Cleanup e2e test endpoints and remove unused screenshot calls (#22674)
* Cleanup e2e test endpoints and remove unused screenshot calls

* Fix lint
