+++ extensions/amp-base-carousel/0.1/test-e2e/test-non-looping.js
@@ -49,7 +49,6 @@ describes.endtoend(
 
         await expect(prop(el, 'scrollWidth')).to.equal(pageWidth * SLIDE_COUNT);
         await waitForCarouselImg(controller, 0);
-        await controller.takeScreenshot('screenshots/render.png');
       });
 
     it('should layout the adjacent slide', async () => {
@@ -74,7 +73,6 @@ describes.endtoend(
       // We should have snapped to the edge of the slide rather than the
       // requested scroll position.
       await expect(prop(el, 'scrollLeft')).to.equal(snappedScrollLeft);
-      await controller.takeScreenshot('screenshots/snapped.png');
     });
 
     //TODO(sparhami): fails on shadow demo
@@ -110,7 +108,6 @@ describes.endtoend(
             'x': 0,
             'width': 900,
           });
-          await controller.takeScreenshot('screenshots/after-resize.png');
         }
       );
   }
