+++ extensions/amp-base-carousel/0.1/test-e2e/test-vertical.js
@@ -53,7 +53,6 @@ describes.endtoend(
       // await expect(prop(el, 'scrollHeight'))
       //     .to.equal(carouselHeight * (2 * (SLIDE_COUNT - 1) + 1));
       await waitForCarouselImg(controller, 0);
-      await controller.takeScreenshot('screenshots/vertical/render.png');
     });
 
     // TODO(sparhami): unskip
@@ -83,7 +82,6 @@ describes.endtoend(
       // We should have snapped to the edge of the slide rather than the
       // requested scroll position.
       await expect(rect(firstSlide)).to.include({y: carouselTop - slideHeight});
-      await controller.takeScreenshot('screenshots/vertical/snapped.png');
     });
 
     describe('looping', () => {
@@ -113,9 +111,6 @@ describes.endtoend(
           y: carouselTop,
           height: slideHeight,
         });
-        await controller.takeScreenshot(
-          'screenshots/vertical/loop-move-forwards-to-end.png'
-        );
       });
 
       // When resting the first few slides should be translated to the bottom.
@@ -144,9 +139,6 @@ describes.endtoend(
           y: carouselTop,
           height: slideHeight,
         });
-        await controller.takeScreenshot(
-          'screenshots/vertical/loop-move-backwards-to-second.png'
-        );
       });
     });
   }
