+++ test/fixtures/e2e/amp-bind/bind-form.html
@@ -15,7 +15,7 @@
   <p id="xhrText" [text]="result.name"></p>
 
   <form method="post"
-      action-xhr="http://localhost:__TEST_SERVER_PORT__/form/post/success" target="_blank"
+      action-xhr="http://localhost:__TEST_SERVER_PORT__/test/form/post/success" target="_blank"
       on="submit-success:AMP.setState({result: event.response})">
     <input id="submitButton" type="submit" value="Submit">
