+++ test/fixtures/e2e/amp-date-picker/blocked-dates.html
@@ -18,14 +18,14 @@
       type="range"
       number-of-months="2"
       allow-blocked-end-date
-      src="http://localhost:__TEST_SERVER_PORT__/date-picker/config.json">
+      src="http://localhost:__TEST_SERVER_PORT__/test/date-picker/config.json">
   </amp-date-picker>
   <amp-date-picker
       id="disallow-blocked"
       layout="fixed-height" height="360"
       type="range"
       number-of-months="2"
-      src="http://localhost:__TEST_SERVER_PORT__/date-picker/config.json">
+      src="http://localhost:__TEST_SERVER_PORT__/test/date-picker/config.json">
   </amp-date-picker>
 </body>
 </html>
